Separate steady-state monitor from loadgen
