+++ tools/steady-state-monitor/src/main/java/com/slack/astra/tools/monitor/Env.java
@@ -1,4 +1,4 @@
-package com.slack.astra.tools.loadgen;
+package com.slack.astra.tools.monitor;
 
 final class Env {
   private Env() {}
