+++ tools/steady-state-monitor/src/main/java/com/slack/astra/tools/monitor/MonitorMain.java
@@ -1,0 +1,7 @@
+package com.slack.astra.tools.monitor;
+
+public class MonitorMain {
+  public static void main(String[] args) throws Exception {
+    SteadyStateMonitor.runFromEnvironment();
+  }
+}
